jsx > js, match layout branch format
