+++ app/(tabs)/_layout.js
@@ -1,0 +1,15 @@
+import { Stack } from "expo-router";
+
+export default function RootLayoutNav() {
+  return (
+    <>
+      <Stack
+        screenOptions={{
+          headerShown: false,
+        }}
+      >
+        <Stack.Screen name="index" />
+      </Stack>
+    </>
+  );
+}
