--- app/_layout.jsx
@@ -1,68 +1,0 @@
-import FontAwesome from "@expo/vector-icons/FontAwesome";
-import * as Font from "expo-font";
-import { SplashScreen, Stack } from "expo-router";
-import { useEffect, useState, useCallback } from "react";
-import { Provider } from "../context/Auth";
-
-export {
-  // Catch any errors thrown by the Layout component.
-  ErrorBoundary,
-} from "expo-router";
-
-export const unstable_settings = {
-  // Ensure that reloading on `/modal` keeps a back button present.
-  initialRouteName: "(tabs)",
-};
-
-export default function RootLayout() {
-  const [appIsReady, setAppIsReady] = useState(false);
-  useEffect(() => {
-    async function prepare() {
-      try {
-        // Pre-load fonts, make any API calls you need to do here
-        await Font.loadAsync({
-          SpaceMono: require("../assets/fonts/SpaceMono-Regular.ttf"),
-          Karma: require("../assets/fonts/Karma-Regular.ttf"),
-          ...FontAwesome.font,
-        });
-      } catch (e) {
-        console.warn(e);
-      } finally {
-        // Tell the application to render
-        setAppIsReady(true);
-      }
-    }
-
-    prepare();
-  }, []);
-
-  const onLayoutRootView = useCallback(async () => {
-    if (appIsReady) {
-      // This tells the splash screen to hide immediately! If we call this after
-      // `setAppIsReady`, then we may see a blank screen while the app is
-      // loading its initial state and rendering its first pixels. So instead,
-      // we hide the splash screen once we know the root view has already
-      // performed layout.
-      await SplashScreen.hideAsync();
-    }
-  }, [appIsReady]);
-
-  if (!appIsReady) {
-    return null;
-  }
-
-  return (
-    <>
-      <Provider onLayout={onLayoutRootView}>
-        <Stack
-          screenOptions={{
-            headerShown: false,
-          }}
-        >
-          <Stack.Screen name="(tabs)" />
-          <Stack.Screen name="(auth)" />
-        </Stack>
-      </Provider>
-    </>
-  );
-}
